refactor: test파일 위치 변경
